farmers view and quantity migration fix

+++ database/migrations/2019_06_29_180553_create_purchasetemp_items_table.php
@@ -21,7 +21,7 @@ class CreatePurchasetempItemsTable extends Migration
             $table->decimal('cost_price',15,2);
             $table->decimal('discount',15,2);
             $table->bigInteger('unit_id')->unsigned()->index();
-            $table->integer('quantity');
+            $table->decimal('quantity',8,2);
             $table->decimal('total_cost',15,2);
             $table->timestamps();
